fixed convertion

+++ paralleldomain/decoding/dgp_decoder.py
@@ -245,15 +245,15 @@ class DGPDecoder(Decoder):
     def decode_semantic_segmentation_2d(self, scene_name: str, annotation_identifier: str) -> np.ndarray:
         annotation_path = self._dataset_path / scene_name / annotation_identifier
         with annotation_path.open(mode="rb") as cloud_binary:
-            image_data = np.asarray(imageio.imread(cast(BinaryIO, cloud_binary), format="png"))
+            image_data = np.asarray(imageio.imread(cast(BinaryIO, cloud_binary), format="png")).astype(np.int)
 
             class_ids = (image_data[..., 2:3] << 16) + (image_data[..., 1:2] << 8) + image_data[..., 0:1]
-            return class_ids.astype(np.int)
+            return class_ids
 
     def decode_optical_flow(self, scene_name: str, annotation_identifier: str) -> np.ndarray:
         annotation_path = self._dataset_path / scene_name / annotation_identifier
         with annotation_path.open(mode="rb") as cloud_binary:
-            image_data = np.asarray(imageio.imread(cast(BinaryIO, cloud_binary), format="png"))
+            image_data = np.asarray(imageio.imread(cast(BinaryIO, cloud_binary), format="png")).astype(np.int)
             vectors = (image_data[..., [0, 2]] << 8) + image_data[..., [1, 3]]
 
             return vectors
@@ -261,10 +261,10 @@ class DGPDecoder(Decoder):
     def decode_instance_segmentation_2d(self, scene_name: str, annotation_identifier: str) -> np.ndarray:
         annotation_path = self._dataset_path / scene_name / annotation_identifier
         with annotation_path.open(mode="rb") as cloud_binary:
-            image_data = np.asarray(imageio.imread(cast(BinaryIO, cloud_binary), format="png"))
+            image_data = np.asarray(imageio.imread(cast(BinaryIO, cloud_binary), format="png")).astype(np.int)
 
             instance_ids = (image_data[..., 2:3] << 16) + (image_data[..., 1:2] << 8) + image_data[..., 0:1]
-            return instance_ids.astype(np.int)
+            return instance_ids
 
     def decode_point_cloud(self, scene_name: str, cloud_identifier: str) -> np.ndarray:
         cloud_path = self._dataset_path / scene_name / cloud_identifier
